move the resource using tests to a single file to help windows testers
